Only include the latest version of each source package
Signed-off-by: Adam D. Barratt <adam@adam-barratt.org.uk>

The Sources file for unstable will contain multiple records for a single
source package when one or more architectures has out-of-date binaries.
As the latest version is that which will be considered for migration, only
that version should be added to the list of available sources.

This is not currently an issue in live use, as the data files are already
pre-processed by britney before being passed to britney2.