Add implicit unblocks for t-p-u approvals
Signed-off-by: Adam D. Barratt <adam@adam-barratt.org.uk>

t-p-u approval previously required both an "unblock" and an "approve"
hint if the package was blocked; the "approve" should be sufficient and
this is the simplest method of achieving that.

There are some cases where this does not quite do the right thing (e.g.
for a package which has both a t-p-u "approve" hint and an "unblock"
hint for the package in unstable) but it is preferable to requiring
t-p-u hints to be added in pairs always.