Drop boottest support
This was a giant copy&paste, was disabled four months ago, and the
infrastructure for this ceased to exist.

If this comes back, the AutoPackageTest class should be generalized to also
issue phone boot tests (exposed as new architectures, which should then be
called "platforms"), to avoid all this duplicated code.