Trigger linux rebuild test on upload of gcc-defaults
The update of gcc to gcc-9 introduced a regression in buildability of
anything relying on kernel headers.  This could have been caught by the
kernel's standard rebuild autopkgtest, but we currently only trigger the
linux autopkgtest for source packages named gcc-N, which excludes
gcc-defaults.

Include gcc-defaults in the list of packages that trigger a linux rebuild
test.

Bug-Ubuntu: https://bugs.launchpad.net/bugs/1836100