test_autopkgtest: Strip submit-time
Testsuite regression from 4c27eadffa01bc1c084e947a7fa1e8a7509d1653.

submit-time is the time the request was submitted to amqp. For tests
it's the time the testcase was run. Many of our testcases test the exact
parameters that were passed to amqp, so they now fail because
submit-time is in there but it's not being checked for. It would be
possible to either verify the submit-time too, or switch to fuzzy
matching - but it's actually not very valuable to test submit-time,
since it is only for human readers of autopkgtest-cloud log files or
autopkgtest.ubuntu.com/running.

Instead strip the field from the amqp parameters before any tests see
them.