DependsPolicy: print some info when skipping the dep check
The autopkgtest policy prints when it doesn't run a test due to a
package being unsatisfiable on an arch. This can be a bit mysterious,
and it can seem like it's causing the package to not become a candidate,
when it's not. Print some info when we decide not to skip indicating
what happened.