Patch fetch_db to handle missing 'content-length' header
We don't know why the autopkgtest webserver has stopped providing a
Content-Length header but the current code doesn't handle its absence, so
detect this rather than throwing an exception.

Should be reverted once the header is back, we don't want to be in the dark
about short reads of the db.